[Util] Edit Packaging
