+++ features/feature-home/src/main/java/com/example/presentation/ui_component/ScheduleView.kt
@@ -46,10 +46,10 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.core.model.Schedule
-import com.example.ui_component.DefaultItem
-import com.example.ui_component.DefaultListView
+import com.example.ui_component.template.DefaultItem
+import com.example.ui_component.template.DefaultListView
 import com.example.ui_component.R
-import com.example.ui_component.RoundedIconButton
+import com.example.ui_component.buttons.RoundedIconButton
 import com.example.ui_component.VerticalSpacer
 import com.example.ui_component.values.darkGray
 import com.example.ui_component.values.horizontalGradation
