+++ features/feature-join/src/main/java/com/example/feature_join/presentation/screen/ProfileSettingScreen.kt
@@ -14,7 +14,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -27,7 +26,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.feature_join.presentation.ui_component.InputView_OnlyNum
 import com.example.feature_join.presentation.ui_component.SelectionView
 import com.example.feature_join.presentation.viewmodel.JoinViewModel
-import com.example.ui_component.CustomGradientButton_1
+import com.example.ui_component.buttons.CustomGradientButton_1
 import com.example.ui_component.R
 import com.example.ui_component.values.mainTheme
