+++ features/feature-mypage/src/main/java/com/example/feature_mypage/presentation/ui_component/ModifyProfileImageView.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.ui_component.DefaultEmblemSelectIconView
+import com.example.ui_component.template.DefaultEmblemSelectIconView
 
 @Composable
 fun ModifyProfileImageView(
