+++ features/feature-schedule/src/main/java/com/example/feature_schedule/MakeScheduleScreen.kt
@@ -1,0 +1,8 @@
+package com.example.feature_schedule
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MakeScheduleScreen() {
+
+}
