--- ui-component/src/main/java/com/example/ui_component/TestComponent.kt
@@ -1,9 +1,0 @@
-package com.example.ui_component
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun TestComponent(text: String) {
-    Text(text = text)
-}
