+++ ui-component/src/main/java/com/example/ui_component/buttons/GradientButton.kt
@@ -1,4 +1,4 @@
-package com.example.ui_component
+package com.example.ui_component.buttons
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
