+++ ui-component/src/main/java/com/example/ui_component/template/DefaultEmblemSelectIconView.kt
@@ -1,4 +1,4 @@
-package com.example.ui_component
+package com.example.ui_component.template
 
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -24,6 +24,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
+import com.example.ui_component.R
 
 @Composable
 fun DefaultEmblemSelectIconView(
